server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
supabase==2.9.1
python-dotenv==1.0.1
pyjwt==2.9.0
httpx==0.27.0
openpyxl==3.1.5
